refactor(web): extract WorkflowSubmissionService + SweepSubmissionService (#167)
Move the largest remaining handler (``run_workflow`` @ 340 LOC) and its
helpers out of the router:

- ``WorkflowSubmissionService.run`` — non-sweep path: resolve profile +
  render workflow + hold per-mount sync lock + BFS submit + seed
  transition + open workflow_run watch (+ optional subscription).
- ``SweepSubmissionService.dispatch`` — materialize cells synchronously
  + spawn orchestrator as a background task.
- ``services/_submission_common.py`` gains the Mount-session context
  manager (now raising typed ``MountSyncFailedError`` instead of
  surfacing ``HTTPException(502)`` inline — cleaner flow control in the
  non-sweep ``_fail`` path).

Test patch surface preserved via router module-level attributes:

- ``SweepSpec`` / ``SweepOrchestrator`` / ``WorkflowRunner`` /
  ``SlurmSSHExecutorPool`` stay imported at router scope and are
  listed in ``__all__``; services receive them as constructor args so
  ``unittest.mock.patch('srunx.web.routers.workflows.<Name>')``
  replacements reach the actual call sites.
- ``_dispatch_sweep`` / ``_run_sweep_background`` kept as thin
  router-level shims so ``tests/sweep/test_sweep_ssh_integration.py``'s
  direct ``wf_mod._dispatch_sweep(...)`` calls keep working.

Router drops from 1863 → 322 LOC. All lazy imports (observability
storage repositories, runtime.submission_plan, security.*,
ssh.core.config, config/sync service) moved with their code, none
hoisted. Transaction scoping (``with transaction(conn, 'IMMEDIATE')``)
and python:-prefix guards preserved verbatim.
